+++ OwaspPlaywrightTests/Data/ProductsData.cs
@@ -1,13 +1,13 @@
 namespace OwaspPlaywrightTests.Data;
 
-public sealed record Product(string Name, string Description, decimal Price, bool SoldOut);
+public sealed record Product(string Name, string Description, float Price, bool SoldOut);
 
 public static class ProductsData
 {
     public static readonly Product RaspberryJuice = new(
         Name: "Raspberry Juice (1000ml)",
         Description: "Made from blended Raspberry Pi, water and sugar.",
-        Price: 4.99m,
+        Price: 4.99f,
         SoldOut: false
     );
 }
